consensus: move verify::block to its own file
Part of #477.
